refactor: Update error handling
