+++ src/domain/errors/CarError.js
@@ -1,0 +1,7 @@
+import { ClientError } from "./ClientError";
+
+export class CarError extends ClientError {
+  constructor(message) {
+    super("CarError", message);
+  }
+}
