+++ src/domain/errors/ClientError.js
@@ -1,0 +1,6 @@
+export class ClientError extends Error {
+  constructor(name, message) {
+    super(message);
+    this.name = name;
+  }
+}
