+++ src/domain/errors/RaceError.js
@@ -1,0 +1,7 @@
+import { ClientError } from "./ClientError";
+
+export class RaceError extends ClientError {
+  constructor(message) {
+    super("RaceError", message);
+  }
+}
