+++ src/domain/errors/ViewError.js
@@ -1,0 +1,7 @@
+import { ClientError } from "./ClientError";
+
+export class ViewError extends ClientError {
+  constructor(message) {
+    super("ViewError", message);
+  }
+}
