+++ src/domain/errors/index.js
@@ -1,0 +1,3 @@
+export * from "./CarError";
+export * from "./RaceError";
+export * from "./ViewError";
